+++ packages/cli/src/scaffold/cosmos.test.ts
@@ -1,37 +1,35 @@
-import immutable from 'immutable'
-import Scaffold from './'
+import Scaffold from './index'
 import Protocol from '../protocols'
 
-const protocol = new Protocol('near')
+const protocol = new Protocol('cosmos')
 
 const scaffoldOptions = {
   protocol,
-  contract: 'abc.def.near',
-  network: 'near-mainnet',
-  contractName: 'Contract',
+  network: 'cosmoshub-4',
+  contractName: 'CosmosHub',
 }
 
 const scaffold = new Scaffold(scaffoldOptions)
 
-describe('NEAR subgraph scaffolding', () => {
+describe('Cosmos subgraph scaffolding', () => {
   test('Manifest', () => {
     expect(scaffold.generateManifest()).toEqual(`\
 specVersion: 0.0.5
 schema:
   file: ./schema.graphql
 dataSources:
-  - kind: near
-    name: Contract
-    network: near-mainnet
+  - kind: cosmos
+    name: CosmosHub
+    network: cosmoshub-4
     source:
-      account: "abc.def.near"
+      startBlock: 0
     mapping:
       apiVersion: 0.0.5
       language: wasm/assemblyscript
       entities:
         - ExampleEntity
-      receiptHandlers:
-        - handler: handleReceipt
+      blockHandlers:
+        - handler: handleBlock
       file: ./src/contract.ts
 `)
   })
@@ -48,20 +46,18 @@ type ExampleEntity @entity {
 
   test('Mapping (default)', () => {
     expect(scaffold.generateMapping()).toEqual(`\
-import { near, BigInt } from "@graphprotocol/graph-ts"
+import { cosmos, BigInt } from "@graphprotocol/graph-ts"
 import { ExampleEntity } from "../generated/schema"
 
-export function handleReceipt(
-  receiptWithOutcome: near.ReceiptWithOutcome
-): void {
+export function handleBlock(block: cosmos.Block): void {
   // Entities can be loaded from the store using a string ID; this ID
   // needs to be unique across all entities of the same type
-  let entity = ExampleEntity.load(receiptWithOutcome.receipt.id.toHex())
+  let entity = ExampleEntity.load(block.header.hash.toHex())
 
   // Entities only exist after they have been saved to the store;
   // \`null\` checks allow to create entities on demand
   if (!entity) {
-    entity = new ExampleEntity(receiptWithOutcome.receipt.id.toHex())
+    entity = new ExampleEntity(block.header.hash.toHex())
 
     // Entity fields can be set using simple assignments
     entity.count = BigInt.fromI32(0)
@@ -71,7 +67,7 @@ export function handleReceipt(
   entity.count = entity.count + BigInt.fromI32(1)
 
   // Entity fields can be set based on receipt information
-  entity.block = receiptWithOutcome.block.header.hash
+  entity.height = block.header.height
 
   // Entities can be written to the store with \`.save()\`
   entity.save()
